requirements.txt
amqp==5.1.1
asgiref==3.6.0
async-timeout==4.0.2
billiard==3.6.4.0
bleach==6.0.0
bokeh==2.4.3
cached-property==1.5.2
celery==5.2.7
celery-progress==0.1.3
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.0.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
dj-database-url==1.2.0
Django==3.2.16
django-celery-results==2.4.0
django-crispy-forms==1.14.0
django-filter==22.1
django-multi-form-view==2.0.1
django-tables2==2.5.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==4.13.0
Jinja2==3.1.2
kombu==5.2.4
Markdown==3.4.1
MarkupSafe==2.1.2
numpy==1.21.6
packaging==23.0
pandas==1.3.5
panel==0.14.3
panel-highcharts==20221018.1
param==1.12.3
Pillow==9.4.0
prompt-toolkit==3.0.36
psycopg2-binary==2.9.5
pycparser==2.21
pyct==0.5.0
python-dateutil==2.8.2
pytz==2022.7
pyviz-comms==2.2.1
PyYAML==6.0
rabbitmq==0.2.0
redis==4.4.0
requests==2.28.2
six==1.16.0
sqlparse==0.4.3
tornado==6.2
tqdm==4.64.1
typing_extensions==4.4.0
urllib3==1.26.14
vine==5.0.0
wcwidth==0.2.5
webencodings==0.5.1
whitenoise==6.3.0
zipp==3.11.0
sierralocal==0.2.1
xhtml2pdf==0.2.11
django-xhtml2pdf==0.0.4

